Finished eyes + some decor
As the title says. Finalized all eye-part assets and started on some decor.
Note: Irises now all have the same design, just different colours. Realized the designs were influenced by the colour, which is.. not ideal for this project.
--->> Irises need the iris_blank sprite as an overlay. I can also export it already merged

I used the eye shop as a background while drawing the decor, so hopefully it should fit in size-wise just fine